Refactor as per PR suggestions: * Separate classes into different files * Reduce duplicated code * Nit: rename vars so they make more sense
