+++ api/index.js
@@ -1,4 +1,5 @@
-import { Leetcode, LeetcodeData } from "./leetcode.js";
+import { Leetcode } from "./leetcode.js";
+import { LeetcodeData  } from "./leetcode-data.js";
 import { Stats } from "./stats.js";
 import { getDiscordClient } from "./client.js";
